Robogames day 1.
Added Magic waypoints for tabletop chllenge. That was essially a line following.
Did that by generalizing cone identifier. It is much better structured but needs a lot of work.
Its probably my path to a genral object identifier.

Somewhere in recent changes, bot has developed a crazy oversteer (round and round it goes).
This includes some testing / tracing to identify and correct so I can follow waypoints on day 2.