This is a major re-write of vnavs_mqtt.py SocketWrapper. I was having trouble getting file server to work smoothly. As I dug into that, I realized the the origianl code was pretty hokey. I have restructured to a more rational class organization and added a bunch of comments.
This is totally untested in this commit. Some of the comments ae not
correct because they are transitional notes as I figured this all out.
It also needs more work to handle connect / disconnect / restart
smoothly. It was hard to figure that out before. Its probably easier
now because the structure is more rational.