Get OpticChiasm.ReflexEntities working.
That ended up leading to a signifcant change to make debugging easier. Camerman now has a driver that works with the macbook camera.
That probably works with most USB cameras because it uses Open.

Added an argument processor and bits of supporting code to make it easier to make temporary configuration changes for testing.